It works! Using ncurses
Updated to use ncurses to take keyboard inputs and improve printing the board to the screen.

The game works! The snake is implemented as a doubly linked list, and "moving forwards" involves popping the tail node and prepending it to the front of the list, updating its coordinates in respect to the direction of movement.